fix(food): split errors.ts to satisfy oxlint max-lines on CI
CI's oxlint treats max-lines as error and exits 1 (local oxlint exits 0
but reports the same diagnostic — inconsistent behaviour). Splitting
errors.ts into errors.ts + errors-plan.ts + errors-ingest.ts keeps
each file under the 200-line cap; the public surface is preserved via
re-exports from errors.ts.
